+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added method to get viewer: Wip::User.viewer
+
 ## [1.1.0] - 2020-11-11
+
+### Changed
 
 - Renamed command from wipco to wip
 - Renamed gem from wipco to wip-cli